save repo in magic for reuse in many classes
Helps get rid of repo argument in some methods. It's
annoying to pass it all the time.

More importantly some of these classes have pointer
to git_repository that is not ref-counted, so if
repo gets freed before all pointers to it then some
calls may result in bad access errors or segfaults.