is_foo methods don't require 'require Git::Raw::xxxx’.
Since 'is_foo' is a method of perl, it has an requirements 'require Git::Raw::Blob' or 'require  Git::Raw::Tree' in advance according to the type of an objects.

Since it calls in order that the users of is_foo may identify of which objects it is a type, according to the type of an objects, an import interface is not appropriate in a modules beforehand.